factor launchmon dependencies into a sub-class of STAT_Frontend
